Move zero G roll TEDs to their own file
